controller: Extract firewall to its own mod
